split linkcommand classes
