+++ build.gradle
@@ -26,7 +26,7 @@ repositories {
 }
 
 checkstyle {
-    toolVersion = '8.1'
+    toolVersion = '8.23'
 }
 
 jacocoTestReport {
@@ -60,7 +60,7 @@ dependencies {
 
     implementation group: 'com.fasterxml.jackson.core', name: 'jackson-databind', version: '2.7.0'
     implementation group: 'com.fasterxml.jackson.datatype', name: 'jackson-datatype-jsr310', version: '2.7.4'
-    
+
     implementation group: 'commons-validator', name: 'commons-validator', version: '1.4.0'
 
     testImplementation group: 'org.junit.jupiter', name: 'junit-jupiter-api', version: jUnitVersion
