+++ src/main/java/seedu/module/logic/parser/ModuleBookParser.java
@@ -12,7 +12,7 @@ import seedu.module.logic.commands.DeleteCommand;
 import seedu.module.logic.commands.ExitCommand;
 import seedu.module.logic.commands.FindCommand;
 import seedu.module.logic.commands.HelpCommand;
-import seedu.module.logic.commands.LinkCommand;
+import seedu.module.logic.commands.linkcommands.LinkCommand;
 import seedu.module.logic.commands.ListCommand;
 import seedu.module.logic.parser.exceptions.ParseException;
